Add dev environment for vscode
I added a devcontainer for vscode. These are a great way to baseline
a development environment across multiple users. It will make
development on this project easier as long as you use vscode. If
you do not... then good luck. I'm not supporting it for now. One
day a better IDE will show up and I'll change this.